fix(transform): set standings int columns

+++ transform/models/staging/stg_standings_regular_goals.sql
@@ -1,35 +1,36 @@
 with games as (
 
     select * from {{ ref("stg_games") }}
-	where season_type = 2 -- regular season
+    where season_type = 2 -- regular season
 
 ),
 
 goals as (
-	select 
-		home_team_id as team_id,
-		season, 
-		home_team_score as goals_for,
-		away_team_score as goals_against
-	from games
+    select 
+        home_team_id as team_id,
+        season, 
+        home_team_score as goals_for,
+        away_team_score as goals_against
+    from games
 	
-	union all
+    union all
 	
-	select
-		away_team_id as team_id,
-		season,
-		away_team_score as goals_for,
-		home_team_score  as goals_against
+    select
+        away_team_id as team_id,
+        season,
+        away_team_score as goals_for,
+        home_team_score  as goals_against
 		
-	from games
+    from games
 )
 
 select 
     team_id,
     season,
-    sum(goals_for) as goals_for,
-    sum(goals_against) as goals_against,
-    sum(goals_for) - sum(goals_against) as goals_diff
+    sum(goals_for)::int as goals_for,
+    sum(goals_against)::int as goals_against,
+    (sum(goals_for) - sum(goals_against))::int as goals_diff
 
 from goals
+
 group by 1, 2
